application/moviepedia/src/androidTest/java/org/videolan/moviepedia/database/ConvertersTest.kt
/*******************************************************************************
 *  ConvertersTest.kt
 * ****************************************************************************
 * Copyright © 2018 VLC authors and VideoLAN
 *
 * This program is free software; you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation; either version 2 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program; if not, write to the Free Software
 * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston MA 02110-1301, USA.
 ******************************************************************************/

package org.videolan.moviepedia.database

import androidx.core.net.toUri
import androidx.test.ext.junit.runners.AndroidJUnit4
import junit.framework.Assert.assertEquals
import org.junit.Test
import org.junit.runner.RunWith

@RunWith(AndroidJUnit4::class)
class ConvertersTest {
    private val str = "upnp://http://[fe80::61a1:a5a4:c66:bc5d]:2869/u"
    private val uri = str.toUri()

    @Test fun uriToString() {
        assertEquals(str, Converters().uriToString(uri))
    }

    @Test fun stringToUri() {
        assertEquals(uri, Converters().stringToUri(str))
    }
}